requirements.txt
python-dotenv
langchain @ git+https://github.com/michaeltansg/langchain@3563f13aa8666a526a74ae259003015cf16e424e
openai==0.27.8
tiktoken==0.4.0
atlassian-python-api==3.39.0
tenacity==8.2.2
beautifulsoup4==4.12.2
lxml==4.9.2